Track updates in forced abbreviation
